fix: ExaFMM Windows/MSVC compatibility
- Add NOMINMAX macro to prevent Windows min/max conflicts
- Replace posix_memalign with _aligned_malloc/_aligned_free on Windows
- Fix OpenMP loop indices from size_t to int (MSVC requirement)
- Add EXAFMM_NO_OMP_TASKS macro to disable OpenMP task pragma on MSVC
- Complete MKL DFTI wrapper implementation with all FFT functions
- Add complex_t template specializations for FMM scale-invariant class
- Convert timer.h from gettimeofday to C++ chrono
- Add FFTW constant definitions (FFTW_FORWARD, etc.)

This enables ExaFMM FMM acceleration to build on Windows with MSVC.

+++ external/exafmm-t/include/build_list.h
@@ -256,7 +256,7 @@ namespace exafmm_t {
     unordered_map<uint64_t, size_t> key2id = get_key2id(nodes);
     unordered_set<uint64_t> leaf_keys = get_leaf_keys(nodes);
 #pragma omp parallel for schedule(dynamic)
-    for (size_t i=0; i<nodes.size(); i++) {
+    for (int i=0; i<static_cast<int>(nodes.size()); i++) {
       Node<T>* node = &nodes[i];
       build_M2L_list(node, nodes, key2id);
       build_other_list(node, nodes, fmm, leaf_keys, key2id);
